[clean]関数の書き出し
